+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Features
+
+- Add `proguardUuid` option to `SentryOptions` ([#436](https://github.com/getsentry/sentry-kotlin-multiplatform/pull/436))
+  - This will 
+
+
 ## 0.17.1
 
 ### Fixes